+++ internal/api/dto/fixed-deposit.go
@@ -68,7 +68,7 @@ type FixedDepositItemResponse struct {
 	Description        string             `json:"description"`
 	Unit               string             `json:"unit"`
 	Currency           string             `json:"currency"`
-	Amount             float32            `json:"amount"`
+	Amount             float64            `json:"amount"`
 	SerialNumber       string             `json:"serial_number"`
 	DateOfConfiscation *time.Time         `json:"date_of_confiscation"`
 	CaseNumber         string             `json:"case_number"`
@@ -85,7 +85,7 @@ type FixedDepositDispatchResponse struct {
 	Type         DropdownSimple     `json:"type"`
 	Unit         string             `json:"unit"`
 	Currency     string             `json:"currency"`
-	Amount       float32            `json:"amount"`
+	Amount       float64            `json:"amount"`
 	SerialNumber string             `json:"serial_number"`
 	DateOfAction *time.Time         `json:"date_of_action"`
 	Subject      string             `json:"subject"`
