+++ internal/api/dto/inventory.assessments.go
@@ -20,12 +20,12 @@ type BasicInventoryResponseAssessment struct {
 	InventoryID           int            `json:"inventory_id"`
 	EstimatedDuration     int            `json:"estimated_duration"`
 	Active                bool           `json:"active"`
-	ResidualPrice         *float32       `json:"residual_price"`
+	ResidualPrice         *float64       `json:"residual_price"`
 	DepreciationRate      string         `json:"depreciation_rate"`
 	DepreciationType      DropdownSimple `json:"depreciation_type"`
 	UserProfile           DropdownSimple `json:"user_profile"`
-	GrossPriceNew         float32        `json:"gross_price_new"`
-	GrossPriceDifference  float32        `json:"gross_price_difference"`
+	GrossPriceNew         float64        `json:"gross_price_new"`
+	GrossPriceDifference  float64        `json:"gross_price_difference"`
 	DateOfAssessment      *string        `json:"date_of_assessment"`
 	DateOfEndOfAssessment *string        `json:"date_of_end_of_assessment"`
 	CreatedAt             string         `json:"created_at"`
