+++ internal/api/dto/procurement.article.go
@@ -24,12 +24,12 @@ type ProcurementArticleResponseItem struct {
 	PublicProcurement DropdownSimple `json:"public_procurement"`
 	Title             string         `json:"active"`
 	Description       string         `json:"year"`
-	NetPrice          float32        `json:"net_price"`
+	NetPrice          float64        `json:"net_price"`
 	VATPercentage     *string        `json:"vat_percentage"`
 	Manufacturer      string         `json:"manufacturer"`
 	Amount            int            `json:"amount"`
 	TotalAmount       int            `json:"total_amount"`
-	GrossPrice        float32        `json:"gross_price"`
+	GrossPrice        float64        `json:"gross_price"`
 	VisibilityType    int            `json:"visibility_type"`
 	CreatedAt         string         `json:"created_at"`
 	UpdatedAt         string         `json:"updated_at"`
