+++ internal/api/dto/procurement.plan.go
@@ -53,8 +53,8 @@ type ProcurementPlanResponseItem struct {
 	ApprovedRequests    int                            `json:"approved_requests"`
 	RejectedDescription *string                        `json:"rejected_description"`
 	Items               []*ProcurementItemResponseItem `json:"items"`
-	TotalNet            float32                        `json:"total_net"`
-	TotalGross          float32                        `json:"total_gross"`
+	TotalNet            float64                        `json:"total_net"`
+	TotalGross          float64                        `json:"total_gross"`
 	CreatedAt           string                         `json:"created_at"`
 	UpdatedAt           string                         `json:"updated_at"`
 }
