+++ structs/basic_inventory_assessments.struct.go
@@ -8,9 +8,9 @@ type BasicInventoryAssessmentsTypesItem struct {
 	Active               bool     `json:"active"`
 	DepreciationTypeID   int      `json:"depreciation_type_id"`
 	UserProfileID        int      `json:"user_profile_id"`
-	GrossPriceNew        float32  `json:"gross_price_new"`
-	GrossPriceDifference float32  `json:"gross_price_difference"`
-	ResidualPrice        *float32 `json:"residual_price"`
+	GrossPriceNew        float64  `json:"gross_price_new"`
+	GrossPriceDifference float64  `json:"gross_price_difference"`
+	ResidualPrice        *float64 `json:"residual_price"`
 	DateOfAssessment     *string  `json:"date_of_assessment"`
 	CreatedAt            string   `json:"created_at"`
 	UpdatedAt            string   `json:"updated_at"`
