+++ structs/suppliers.struct.go
@@ -9,7 +9,7 @@ type Suppliers struct {
 	Address       string   `json:"address"`
 	Description   string   `json:"description"`
 	FolderID      int      `json:"folder_id"`
-	TaxPercentage float32  `json:"tax_percentage"`
+	TaxPercentage float64  `json:"tax_percentage"`
 	ParentID      *int     `json:"parent_id"`
 	BankAccounts  []string `json:"bank_accounts"`
 }
